refactor: create pathdb package
