+++ core/trie2/trie.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/NethermindEth/juno/core/crypto"
 	"github.com/NethermindEth/juno/core/felt"
-	"github.com/NethermindEth/juno/core/trie2/triedb"
+	"github.com/NethermindEth/juno/core/trie2/triedb/pathdb"
 	"github.com/NethermindEth/juno/core/trie2/trienode"
 	"github.com/NethermindEth/juno/core/trie2/trieutils"
 	"github.com/NethermindEth/juno/db"
@@ -31,7 +31,7 @@ type Trie struct {
 	hashFn crypto.HashFn
 
 	// The underlying database to store and retrieve trie nodes
-	db triedb.TrieDB
+	db pathdb.TrieDB
 
 	// Check if the trie has been committed. Trie is unusable once committed.
 	committed bool
@@ -54,7 +54,7 @@ type TrieID interface {
 
 // Creates a new trie
 func New(id TrieID, height uint8, hashFn crypto.HashFn, txn db.Transaction) (*Trie, error) {
-	database := triedb.New(txn, id.Bucket())
+	database := pathdb.New(txn, id.Bucket())
 	tr := &Trie{
 		owner:      id.Owner(),
 		height:     height,
@@ -80,7 +80,7 @@ func NewEmpty(height uint8, hashFn crypto.HashFn) *Trie {
 		hashFn:     hashFn,
 		root:       nil,
 		nodeTracer: newTracer(),
-		db:         triedb.EmptyDatabase{},
+		db:         pathdb.EmptyDatabase{},
 	}
 }
