+++ core/trie2/triedb/pathdb/types.go
@@ -1,4 +1,4 @@
-package triedb
+package pathdb
 
 type leafType uint8
